Add support for anyOf properties and apply DRY to createSchema (#582)
* add support for anyOf

* Move createResponseSchema and createRequestSchema functions to common module

* Rename createAnyOneOfProperty for consistency

* Fix linter errors

* Use title as tab label and continue creating nodes if properties exist

---------
