+++ packages/docusaurus-plugin-openapi-docs/src/openapi/createRequestExample.ts
@@ -8,7 +8,7 @@
 import chalk from "chalk";
 import merge from "lodash/merge";
 
-import { mergeAllOf } from "../markdown/createRequestSchema";
+import { mergeAllOf } from "../markdown/createSchema";
 import { SchemaObject } from "./types";
 
 interface OASTypeToTypeMap {
