+++ packages/docusaurus-plugin-openapi-docs/src/openapi/createResponseExample.ts
@@ -8,7 +8,7 @@
 import chalk from "chalk";
 import merge from "lodash/merge";
 
-import { mergeAllOf } from "../markdown/createResponseSchema";
+import { mergeAllOf } from "../markdown/createSchema";
 import { SchemaObject } from "./types";
 
 interface OASTypeToTypeMap {
